refactor: uuid 생성기 util로 이동 #1
