+++ second-hand/src/main/java/kr/codesquad/util/UuidGenerator.java
@@ -1,4 +1,4 @@
-package kr.codesquad.common;
+package kr.codesquad.util;
 
 import java.util.UUID;
